+++ FAnsi.Core/Discovery/Queryable/FAnsiQueryProvider.cs
@@ -113,7 +113,7 @@ namespace FAnsi.Discovery.QueryableAbstraction
         /// Executes a query and returns the result.
         /// This is where expression trees are translated to SQL and executed.
         /// </summary>
-        [RequiresDynamicCode()]
+        [RequiresDynamicCode("Calls System.Type.MakeGenericType(params Type[])")]
         public object Execute(Expression expression)
         {
             if (expression == null)